apache-spark/apache-connect/requirements-local.txt
pyspark==3.4.4
pandas>=1.0.5
pyarrow>=1.0.0
grpcio>=1.48.1
protobuf
grpcio-status
